Updated the Introduction to Data Bags - Group Exercise Complete
Completed the group exercise of creating the users cookbook, updating the run list, converging and verifying the node.

There is a new assumption that they have a node already bootstrapped. The node is named node1. This nodes does not already have a run list but it does not hurt if it already has one.